feat: 完善种子排序和能量调度
- 使用工厂设计模式
- 实现4种不同的能量调度方式

+++ src/main/java/com/example/fuzzer/schedule/core/SeedScheduler.java
@@ -8,26 +8,30 @@ import com.example.fuzzer.schedule.model.Seed;
 public interface SeedScheduler {
     /**
      * 选择下一个要执行的种子
+     *
      * @return 下一个种子，如果没有可用种子则返回null
      */
     Seed selectNextSeed();
 
     /**
      * 添加新的种子到调度器
+     *
      * @param seed 要添加的种子
      */
     void addSeed(Seed seed);
 
     /**
      * 更新种子的性能指标
-     * @param data 种子数据
+     *
+     * @param data          种子数据
      * @param executionTime 执行时间
-     * @param newBranches 新发现的分支数
+     * @param newBranches   新发现的分支数
      */
-    void updatePerformance(byte[] data, int executionTime, int newBranches);
+    void updatePerformance(byte[] data, long executionTime, int newBranches);
 
     /**
      * 获取当前队列中种子的数量
+     *
      * @return 队列中种子的数量
      */
     int getQueueSize();
